renamed Editor to Slate since it holds two exports now and updated import paths
