+++ client/components/CreatePost/Form/formbodycontent.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Editor } from "../../Editor";
+import { Editor } from "../../Slate";
 import SubmitFormBodyContentTitle from "./formbodycontenttitle";
 import { useCreatePost } from "../../../contexts/createpost";
