+++ client/components/Slate/utils/verifiers.tsx
