+++ client/sass/components/_slate.scss
