+++ client/sass/main.scss
@@ -7,7 +7,6 @@
 @import "components/button";
 @import "components/checkbox";
 @import "components/createpost";
-@import "components/editor";
 @import "components/faqdropdown";
 @import "components/footer";
 @import "components/grid";
@@ -22,6 +21,7 @@
 @import "components/radio";
 @import "components/recentposts";
 @import "components/requireddot";
+@import "components/slate";
 @import "components/sortview";
 @import "components/staticfooter";
 @import "components/subredditinfo";
